refactor: organize unit tests into command specific modules
