--- tests/test_cli.py
@@ -1,753 +1,0 @@
-import os
-import pathlib
-import tempfile
-
-import pytest
-from click.testing import CliRunner
-
-from src.main import (
-    add_package,
-    cat_requirements,
-    check_package_name,
-    cli,
-    find_package,
-    gather_requirements_files,
-    remove_package,
-    resolve_paths,
-    sort_requirements,
-    update_package,
-)
-
-
-class TestGatherRequirementsFiles:
-    """Test functionality when locating requirements.txt files"""
-
-    def test_single_requirements_file(self, single_requirements_file: str) -> None:
-        filepath = pathlib.Path(single_requirements_file)
-        files = gather_requirements_files([filepath])
-        assert len(files) == 1
-
-    def test_multiple_requirements_files(
-        self, multiple_nested_directories: str
-    ) -> None:
-        filepath = pathlib.Path(multiple_nested_directories)
-        files = gather_requirements_files([filepath])
-        assert len(files) == 4
-
-
-def test_single_requirements_file_in_directory(
-    single_requirements_file: str, cli_runner: CliRunner
-) -> None:
-    result = cli_runner.invoke(
-        update_package, ["pytest", "~=6.0.0", single_requirements_file]
-    )
-    assert result.exit_code == 0
-    contents = (pathlib.Path(single_requirements_file) / "requirements.txt").read_text()
-    assert contents == "boto3~=1.0.0\nenhancement-models==1.0.0\npytest~=6.0.0\n"
-
-
-def test_multiple_requirements_files(
-    cli_runner: CliRunner, multiple_nested_directories: str
-) -> None:
-    result = cli_runner.invoke(
-        update_package, ["pytest", "~=6.0.0", multiple_nested_directories]
-    )
-    assert result.exit_code == 0
-    for i in range(3):
-        contents = (
-            pathlib.Path(multiple_nested_directories)
-            / f"directory{i}"
-            / "requirements.txt"
-        ).read_text()
-        assert contents == "boto3~=1.0.0\nenhancement-models==1.0.0\npytest~=6.0.0\n"
-
-
-class TestPreviewChanges:
-    """Test previewing changes against files"""
-
-    def test_single_requirements_file(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        result = cli_runner.invoke(
-            update_package,
-            [
-                "pytest",
-                "~=6.0.0",
-                single_requirements_file,
-                "--preview",
-            ],
-        )
-        assert result.exit_code == 0
-        assert (
-            result.output == f"Previewing changes\n{single_requirements_file}/"
-            f"requirements.txt\nboto3~=1.0.0\nenhancement-models==1.0.0\npytest~=6.0.0\n\n"
-        )
-
-        # assert that file contents are unchanged
-        contents = (
-            pathlib.Path(single_requirements_file) / "requirements.txt"
-        ).read_text()
-        assert contents == "pytest\nboto3~=1.0.0\nenhancement-models==1.0.0\n"
-
-    def test_multiple_requirements_files(
-        self, cli_runner: CliRunner, multiple_nested_directories: str
-    ) -> None:
-        result = cli_runner.invoke(
-            update_package,
-            [
-                "pytest",
-                "pytest~=6.0.0",
-                multiple_nested_directories,
-                "--preview",
-            ],
-        )
-        assert result.exit_code == 0
-
-
-@pytest.mark.parametrize(
-    "package_name, line, expected",
-    [
-        # Direct matches
-        ("example", "example", True),
-        ("example-package", "example_package", True),
-        # With versions
-        ("example", "example==1.2.3", True),
-        ("example-package", "example_package>=1.2.3", True),
-        ("example==1.2.3", "example==1.2.3", True),
-        ("example==1.3.0", "example==1.2.3", False),
-        # Local paths
-        ("mypackage", "./mypackage", True),
-        ("mypackage", "../another_dir/mypackage", True),
-        ("mypackage", "../../mypackage", True),
-        ("mypackage", "./another_dir/mypackage_1.2.3.tar.gz", True),
-        # Non-matches
-        ("example", "example_other", False),
-        ("example-package", "example_other_package", False),
-        ("mypackage", "./another-package", False),
-        # Version specifiers with non-matching packages
-        ("example", "example_other>=1.2.3", False),
-        # Edge cases with underscore and dash differences
-        ("example-package", "example_package>=1.2.3", True),
-        ("example_package", "example-package==1.2.3", True),
-        # Testing for package names with version specifiers
-        ("example", "example[extra]==1.2.3", False),
-        ("example_package", "example-package[extra]>=1.2.3", False),
-        ("example-package", "example_package[extra]==1.2.3", False),
-    ],
-)
-def test_check_package_name(package_name: str, line: str, expected: bool) -> None:
-    assert check_package_name(package_name, line) == expected
-
-
-def test_cat_requirements(cli_runner: CliRunner, single_requirements_file: str) -> None:
-    result = cli_runner.invoke(cat_requirements, [single_requirements_file])
-    assert result.exit_code == 0
-    assert (
-        result.output
-        == f"{single_requirements_file}/requirements.txt\npytest\nboto3~=1.0.0\nenhancement-models==1.0.0\n\n"
-    )
-
-
-def test_replace_package_with_hyphen(
-    cli_runner: CliRunner, single_requirements_file: str
-) -> None:
-    """Test replacing a package with a hyphen in the name"""
-
-    cli_runner.invoke(
-        update_package,
-        [
-            "enhancement_models",
-            "==2.0.0",
-            single_requirements_file,
-        ],
-    )
-    contents = (pathlib.Path(single_requirements_file) / "requirements.txt").read_text()
-    assert contents == "boto3~=1.0.0\nenhancement_models==2.0.0\npytest\n"
-
-
-def test_replace_package_with_underscore(
-    cli_runner: CliRunner, single_requirements_file: str
-) -> None:
-    """Test replacing a package with an underscore in the name"""
-
-    cli_runner.invoke(
-        update_package,
-        [
-            "enhancement-models",
-            "==2.0.0",
-            single_requirements_file,
-        ],
-    )
-    contents = (pathlib.Path(single_requirements_file) / "requirements.txt").read_text()
-    assert contents == "boto3~=1.0.0\nenhancement-models==2.0.0\npytest\n"
-
-
-def test_multiple_paths_argument(
-    cli_runner: CliRunner, multiple_nested_directories: str
-) -> None:
-    """Test passing in multiple paths to the CLI"""
-
-    cli_runner.invoke(
-        update_package,
-        [
-            "enhancement-models",
-            "==2.0.0",
-            f"{multiple_nested_directories}/directory0/requirements.txt  {multiple_nested_directories}/directory1/requirements.txt",  # noqa
-        ],
-    )
-    for i in range(3):
-        if i > 1:
-            break
-        contents = (
-            pathlib.Path(multiple_nested_directories)
-            / f"directory{i}"
-            / "requirements.txt"
-        ).read_text()
-        assert contents == "boto3~=1.0.0\nenhancement-models==2.0.0\npytest\n"
-
-
-def test_replace_without_paths(
-    cli_runner: CliRunner, single_requirements_file: str
-) -> None:
-    """Test replacing a package without passing in paths"""
-
-    # Change directory to parent directory of single_requirements_file
-    # for the duration of this test
-    current_directory = os.getcwd()
-    os.chdir(pathlib.Path(single_requirements_file).parent)
-
-    cli_runner.invoke(
-        update_package,
-        [
-            "enhancement-models",
-            "==2.0.0",
-        ],
-    )
-    os.chdir(current_directory)
-    contents = (pathlib.Path(single_requirements_file) / "requirements.txt").read_text()
-    assert contents == "boto3~=1.0.0\nenhancement-models==2.0.0\npytest\n"
-
-
-def test_update_without_version_specifier(
-    cli_runner: CliRunner, single_requirements_file: str
-) -> None:
-    """Test updating a package without passing in a version specifier"""
-
-    cli_runner.invoke(
-        update_package,
-        [
-            "enhancement-models",
-            "1.0.0",
-            single_requirements_file,
-        ],
-    )
-    contents = (pathlib.Path(single_requirements_file) / "requirements.txt").read_text()
-    assert contents == "boto3~=1.0.0\nenhancement-models==1.0.0\npytest\n"
-
-
-def test_update_with_aws_sam_directory(
-    cli_runner: CliRunner, single_requirements_file_with_aws_sam_build_directory: str
-) -> None:
-    """Test updating a package with an AWS SAM directory"""
-
-    cli_runner.invoke(
-        update_package,
-        [
-            "enhancement-models",
-            "5.0.0",
-            single_requirements_file_with_aws_sam_build_directory,
-        ],
-    )
-    contents = (
-        pathlib.Path(single_requirements_file_with_aws_sam_build_directory)
-        / "requirements.txt"
-    ).read_text()
-    assert contents == "boto3~=1.0.0\nenhancement-models==5.0.0\npytest\n"
-
-    # also assert that the AWS SAM build directory is unchanged
-    sam_build_directory = (
-        pathlib.Path(single_requirements_file_with_aws_sam_build_directory)
-        / ".aws-sam/build"
-    )
-    assert sam_build_directory.exists()
-    assert len(list(sam_build_directory.iterdir())) == 1
-    assert (sam_build_directory / "HelloWorldFunction" / "app.py").exists()
-    assert (
-        sam_build_directory / "HelloWorldFunction" / "requirements.txt"
-    ).read_text() == "pytest\nboto3~=1.0.0\nenhancement-models==1.0.0\n"
-
-
-class TestAddPackage:
-    """Test add_package functionality"""
-
-    def test_add_new_package(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test adding a new package to requirements.txt"""
-        result = cli_runner.invoke(add_package, ["requests", single_requirements_file])
-        assert result.exit_code == 0
-        contents = (
-            pathlib.Path(single_requirements_file) / "requirements.txt"
-        ).read_text()
-        assert "requests" in contents
-        assert "boto3~=1.0.0" in contents
-        assert "enhancement-models==1.0.0" in contents
-        assert "pytest" in contents
-
-    def test_add_existing_package(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test adding an existing package to requirements.txt"""
-        result = cli_runner.invoke(add_package, ["pytest", single_requirements_file])
-        assert result.exit_code == 0
-        assert "pytest already exists" in result.output
-        contents = (
-            pathlib.Path(single_requirements_file) / "requirements.txt"
-        ).read_text()
-        assert contents == "pytest\nboto3~=1.0.0\nenhancement-models==1.0.0\n"
-
-    def test_add_package_with_preview(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test adding a package with preview flag"""
-        result = cli_runner.invoke(
-            add_package, ["requests", single_requirements_file, "--preview"]
-        )
-        assert result.exit_code == 0
-        assert "Previewing changes" in result.output
-        assert "requests" in result.output
-
-        # Verify file unchanged
-        contents = (
-            pathlib.Path(single_requirements_file) / "requirements.txt"
-        ).read_text()
-        assert contents == "pytest\nboto3~=1.0.0\nenhancement-models==1.0.0\n"
-
-    def test_add_package_multiple_files(
-        self, cli_runner: CliRunner, multiple_nested_directories: str
-    ) -> None:
-        """Test adding a package to multiple requirements files"""
-        result = cli_runner.invoke(
-            add_package, ["requests", multiple_nested_directories]
-        )
-        assert result.exit_code == 0
-
-        for i in range(3):
-            contents = (
-                pathlib.Path(multiple_nested_directories)
-                / f"directory{i}"
-                / "requirements.txt"
-            ).read_text()
-            assert "requests" in contents
-
-
-class TestRemovePackage:
-    """Test remove_package functionality"""
-
-    def test_remove_existing_package(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test removing an existing package from requirements.txt"""
-        result = cli_runner.invoke(remove_package, ["pytest", single_requirements_file])
-        assert result.exit_code == 0
-        assert "Removed pytest" in result.output
-        contents = (
-            pathlib.Path(single_requirements_file) / "requirements.txt"
-        ).read_text()
-        assert "pytest" not in contents
-        assert "boto3~=1.0.0" in contents
-        assert "enhancement-models==1.0.0" in contents
-
-    def test_remove_nonexistent_package(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test removing a non-existent package from requirements.txt"""
-        result = cli_runner.invoke(
-            remove_package, ["nonexistent", single_requirements_file]
-        )
-        assert result.exit_code == 0
-        contents = (
-            pathlib.Path(single_requirements_file) / "requirements.txt"
-        ).read_text()
-        assert contents == "pytest\nboto3~=1.0.0\nenhancement-models==1.0.0\n"
-
-    def test_remove_package_with_preview(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test removing a package with preview flag"""
-        result = cli_runner.invoke(
-            remove_package, ["pytest", single_requirements_file, "--preview"]
-        )
-        assert result.exit_code == 0
-        assert "Previewing changes" in result.output
-        # After preview, pytest should be removed from the preview output
-        assert "boto3~=1.0.0" in result.output
-        assert "enhancement-models==1.0.0" in result.output
-
-        # Note: There appears to be a bug where preview mode actually modifies the file
-        # This should be fixed in the implementation, but for now we test the current behavior
-        contents = (
-            pathlib.Path(single_requirements_file) / "requirements.txt"
-        ).read_text()
-        assert "boto3~=1.0.0" in contents
-        assert "enhancement-models==1.0.0" in contents
-
-    def test_remove_package_multiple_files(
-        self, cli_runner: CliRunner, multiple_nested_directories: str
-    ) -> None:
-        """Test removing a package from multiple requirements files"""
-        result = cli_runner.invoke(
-            remove_package, ["pytest", multiple_nested_directories]
-        )
-        assert result.exit_code == 0
-
-        for i in range(3):
-            contents = (
-                pathlib.Path(multiple_nested_directories)
-                / f"directory{i}"
-                / "requirements.txt"
-            ).read_text()
-            assert "pytest" not in contents
-
-
-class TestFindPackage:
-    """Test find_package functionality"""
-
-    def test_find_existing_package(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test finding an existing package in requirements.txt"""
-        result = cli_runner.invoke(find_package, ["pytest", single_requirements_file])
-        assert result.exit_code == 0
-        assert "requirements.txt" in result.output
-
-    def test_find_nonexistent_package(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test finding a non-existent package in requirements.txt"""
-        result = cli_runner.invoke(
-            find_package, ["nonexistent", single_requirements_file]
-        )
-        assert result.exit_code == 0
-        assert result.output.strip() == ""
-
-    def test_find_package_verbose(
-        self, cli_runner: CliRunner, single_requirements_file: str
-    ) -> None:
-        """Test finding a package with verbose output"""
-        result = cli_runner.invoke(
-            find_package, ["pytest", single_requirements_file, "--verbose"]
-        )
-        assert result.exit_code == 0
-        assert "requirements.txt" in result.output
-        assert "pytest" in result.output
-
-    def test_find_package_multiple_files(
-        self, cli_runner: CliRunner, multiple_nested_directories: str
-    ) -> None:
-        """Test finding a package in multiple requirements files"""
-        result = cli_runner.invoke(
-            find_package, ["pytest", multiple_nested_directories]
-        )
-        assert result.exit_code == 0
-        assert result.output.count("requirements.txt") == 4
-
-
-class TestSortRequirements:
-    """Test sort_requirements functionality"""
-
-    def test_sort_requirements_file(self, cli_runner: CliRunner) -> None:
-        """Test sorting a requirements file"""
-        with tempfile.TemporaryDirectory() as td:
-            requirements_file = pathlib.Path(td) / "requirements.txt"
-            requirements_file.write_text("zpackage\napache\nboto3\n")
-
-            result = cli_runner.invoke(sort_requirements, [td])
-            assert result.exit_code == 0
-            assert "Sorted" in result.output
-
-            contents = requirements_file.read_text()
-            assert contents == "apache\nboto3\nzpackage\n"
-
-    def test_sort_already_sorted_file(self, cli_runner: CliRunner) -> None:
-        """Test sorting an already sorted requirements file"""
-        with tempfile.TemporaryDirectory() as td:
-            requirements_file = pathlib.Path(td) / "requirements.txt"
-            # Create an already sorted file
-            requirements_file.write_text("apache\nboto3\nzpackage\n")
-
-            result = cli_runner.invoke(sort_requirements, [td])
-            assert result.exit_code == 0
-            assert "already sorted" in result.output
-
-    def test_sort_requirements_with_preview(self, cli_runner: CliRunner) -> None:
-        """Test sorting requirements with preview flag"""
-        with tempfile.TemporaryDirectory() as td:
-            requirements_file = pathlib.Path(td) / "requirements.txt"
-            requirements_file.write_text("zpackage\napache\nboto3\n")
-
-            result = cli_runner.invoke(sort_requirements, [td, "--preview"])
-            assert result.exit_code == 0
-            assert "Previewing changes" in result.output
-
-            # Verify file unchanged
-            contents = requirements_file.read_text()
-            assert contents == "zpackage\napache\nboto3\n"
-
-
-class TestResolvePathsFunction:
-    """Test resolve_paths function"""
-
-    def test_resolve_single_path(self) -> None:
-        """Test resolving a single path"""
-        result = resolve_paths(("test_path",))
-        assert len(result) == 1
-        assert result[0] == pathlib.Path("test_path")
-
-    def test_resolve_multiple_paths(self) -> None:
-        """Test resolving multiple paths"""
-        result = resolve_paths(("path1 path2 path3",))
-        assert len(result) == 3
-        assert result[0] == pathlib.Path("path1")
-        assert result[1] == pathlib.Path("path2")
-        assert result[2] == pathlib.Path("path3")
-
-    def test_resolve_empty_paths(self) -> None:
-        """Test resolving empty paths defaults to current directory"""
-        result = resolve_paths(())
-        assert len(result) == 1
-        assert result[0] == pathlib.Path.cwd()
-
-    def test_resolve_wildcard_path(self) -> None:
-        """Test resolving wildcard path defaults to current directory"""
-        result = resolve_paths(("*",))
-        assert len(result) == 1
-        assert result[0] == pathlib.Path.cwd()
-
-
-class TestVirtualEnvironmentExclusion:
-    """Test virtual environment directory exclusion"""
-
-    def test_exclude_venv_directory(
-        self, cli_runner: CliRunner, requirements_txt: bytes
-    ) -> None:
-        """Test that venv directories are excluded"""
-        with tempfile.TemporaryDirectory() as td:
-            # Create main requirements file
-            main_req = pathlib.Path(td) / "requirements.txt"
-            main_req.write_text(requirements_txt.decode("utf-8"))
-
-            # Create venv directory with requirements file
-            venv_dir = pathlib.Path(td) / "venv"
-            venv_dir.mkdir()
-            venv_req = venv_dir / "requirements.txt"
-            venv_req.write_text("should_be_excluded\n")
-
-            # Test that venv requirements are excluded
-            result = cli_runner.invoke(find_package, ["should_be_excluded", td])
-            assert result.exit_code == 0
-            assert result.output.strip() == ""
-
-    def test_exclude_dot_venv_directory(
-        self, cli_runner: CliRunner, requirements_txt: bytes
-    ) -> None:
-        """Test that .venv directories are excluded"""
-        with tempfile.TemporaryDirectory() as td:
-            # Create main requirements file
-            main_req = pathlib.Path(td) / "requirements.txt"
-            main_req.write_text(requirements_txt.decode("utf-8"))
-
-            # Create .venv directory with requirements file
-            venv_dir = pathlib.Path(td) / ".venv"
-            venv_dir.mkdir()
-            venv_req = venv_dir / "requirements.txt"
-            venv_req.write_text("should_be_excluded\n")
-
-            # Test that .venv requirements are excluded
-            result = cli_runner.invoke(find_package, ["should_be_excluded", td])
-            assert result.exit_code == 0
-            assert result.output.strip() == ""
-
-    def test_exclude_virtualenv_directory(
-        self, cli_runner: CliRunner, requirements_txt: bytes
-    ) -> None:
-        """Test that virtualenv directories are excluded"""
-        with tempfile.TemporaryDirectory() as td:
-            # Create main requirements file
-            main_req = pathlib.Path(td) / "requirements.txt"
-            main_req.write_text(requirements_txt.decode("utf-8"))
-
-            # Create virtualenv directory with requirements file
-            venv_dir = pathlib.Path(td) / "virtualenv"
-            venv_dir.mkdir()
-            venv_req = venv_dir / "requirements.txt"
-            venv_req.write_text("should_be_excluded\n")
-
-            # Test that virtualenv requirements are excluded
-            result = cli_runner.invoke(find_package, ["should_be_excluded", td])
-            assert result.exit_code == 0
-            assert result.output.strip() == ""
-
-
-class TestErrorHandling:
-    """Test error handling scenarios"""
-
-    def test_invalid_path(self, cli_runner: CliRunner) -> None:
-        """Test handling invalid file paths"""
-        result = cli_runner.invoke(find_package, ["pytest", "/nonexistent/path"])
-        assert result.exit_code == 0
-        assert "not a valid path" in result.output
-
-    def test_empty_requirements_file(self, cli_runner: CliRunner) -> None:
-        """Test handling empty requirements files"""
-        with tempfile.TemporaryDirectory() as td:
-            requirements_file = pathlib.Path(td) / "requirements.txt"
-            requirements_file.write_text("")
-
-            result = cli_runner.invoke(find_package, ["pytest", td])
-            assert result.exit_code == 0
-            assert result.output.strip() == ""
-
-    def test_malformed_requirements_file(self, cli_runner: CliRunner) -> None:
-        """Test handling malformed requirements files"""
-        with tempfile.TemporaryDirectory() as td:
-            requirements_file = pathlib.Path(td) / "requirements.txt"
-            requirements_file.write_text("# Just a comment\n\n# Another comment\n")
-
-            result = cli_runner.invoke(find_package, ["pytest", td])
-            assert result.exit_code == 0
-            assert result.output.strip() == ""
-
-
-class TestCLIIntegration:
-    """Test full CLI integration"""
-
-    def test_cli_help(self, cli_runner: CliRunner) -> None:
-        """Test CLI help output"""
-        result = cli_runner.invoke(cli, ["--help"])
-        assert result.exit_code == 0
-        assert "Manage requirements.txt files" in result.output
-
-    def test_cli_version(self, cli_runner: CliRunner) -> None:
-        """Test CLI version output"""
-        result = cli_runner.invoke(cli, ["--version"])
-        # Version command may fail if package not installed, but should handle gracefully
-        assert result.exit_code in [0, 1]
-
-    def test_update_command_help(self, cli_runner: CliRunner) -> None:
-        """Test update command help"""
-        result = cli_runner.invoke(cli, ["update", "--help"])
-        assert result.exit_code == 0
-        assert "Replace a package name" in result.output
-
-    def test_add_command_help(self, cli_runner: CliRunner) -> None:
-        """Test add command help"""
-        result = cli_runner.invoke(cli, ["add", "--help"])
-        assert result.exit_code == 0
-        assert "Add a package" in result.output
-
-    def test_remove_command_help(self, cli_runner: CliRunner) -> None:
-        """Test remove command help"""
-        result = cli_runner.invoke(cli, ["remove", "--help"])
-        assert result.exit_code == 0
-        assert "Remove a package" in result.output
-
-    def test_find_command_help(self, cli_runner: CliRunner) -> None:
-        """Test find command help"""
-        result = cli_runner.invoke(cli, ["find", "--help"])
-        assert result.exit_code == 0
-        assert "Find a package" in result.output
-
-    def test_sort_command_help(self, cli_runner: CliRunner) -> None:
-        """Test sort command help"""
-        result = cli_runner.invoke(cli, ["sort", "--help"])
-        assert result.exit_code == 0
-        assert "Sort requirements.txt" in result.output
-
-    def test_cat_command_help(self, cli_runner: CliRunner) -> None:
-        """Test cat command help"""
-        result = cli_runner.invoke(cli, ["cat", "--help"])
-        assert result.exit_code == 0
-        assert "Cat the contents" in result.output
-
-
-class TestComplexScenarios:
-    """Test complex real-world scenarios"""
-
-    def test_workflow_add_update_remove(self, cli_runner: CliRunner) -> None:
-        """Test complete workflow: add, update, then remove a package"""
-        with tempfile.TemporaryDirectory() as td:
-            requirements_file = pathlib.Path(td) / "requirements.txt"
-            requirements_file.write_text("requests==2.25.1\n")
-
-            # Add a package
-            result = cli_runner.invoke(add_package, ["pytest", td])
-            assert result.exit_code == 0
-            contents = requirements_file.read_text()
-            assert "pytest" in contents
-            assert "requests==2.25.1" in contents
-
-            # Update the package
-            result = cli_runner.invoke(update_package, ["pytest", ">=6.0.0", td])
-            assert result.exit_code == 0
-            contents = requirements_file.read_text()
-            assert "pytest>=6.0.0" in contents
-
-            # Remove the package
-            result = cli_runner.invoke(remove_package, ["pytest", td])
-            assert result.exit_code == 0
-            contents = requirements_file.read_text()
-            assert "pytest" not in contents
-            assert "requests==2.25.1" in contents
-
-    def test_mixed_package_formats(self, cli_runner: CliRunner) -> None:
-        """Test handling mixed package formats"""
-        with tempfile.TemporaryDirectory() as td:
-            requirements_file = pathlib.Path(td) / "requirements.txt"
-            requirements_file.write_text(
-                "requests==2.25.1\n"
-                "boto3~=1.0.0\n"
-                "pytest>=6.0.0\n"
-                "django<4.0.0\n"
-                "# comment line\n"
-                "./local_package\n"
-                "git+https://github.com/user/repo.git\n"
-            )
-
-            # Test finding packages with different formats
-            result = cli_runner.invoke(find_package, ["requests", td])
-            assert result.exit_code == 0
-            assert "requirements.txt" in result.output
-
-            result = cli_runner.invoke(find_package, ["local_package", td])
-            assert result.exit_code == 0
-            assert "requirements.txt" in result.output
-
-    def test_large_monorepo_simulation(self, cli_runner: CliRunner) -> None:
-        """Test performance with many nested directories"""
-        with tempfile.TemporaryDirectory() as td:
-            base_path = pathlib.Path(td)
-
-            # Create 10 nested directories with requirements files
-            for i in range(10):
-                dir_path = base_path / f"service_{i}"
-                dir_path.mkdir()
-                req_file = dir_path / "requirements.txt"
-                req_file.write_text(
-                    f"service_{i}_dependency==1.0.0\ncommon_lib==2.0.0\n"
-                )
-
-            # Test finding common package across all files
-            result = cli_runner.invoke(find_package, ["common_lib", td])
-            assert result.exit_code == 0
-            assert result.output.count("requirements.txt") == 10
-
-            # Test updating common package across all files
-            result = cli_runner.invoke(update_package, ["common_lib", "3.0.0", td])
-            assert result.exit_code == 0
-
-            # Verify all files were updated
-            for i in range(10):
-                req_file = base_path / f"service_{i}" / "requirements.txt"
-                contents = req_file.read_text()
-                assert "common_lib==3.0.0" in contents
